UPDATE
Adjustment at the UPlaythrough and UCharismaAPI, moving functions from Playthrough to API, changing them to static (As requested at Discord).
